+++ app/breakpoints.ts
@@ -1,0 +1,10 @@
+export default {
+    mobile: '(max-width: 767px)',
+    tablet: '(min-width: 768px) and (max-width: 991px)',
+    desktop: '(min-width: 992px)',
+    // Defaults are below. Currently, the OSF does not differentiate
+    // between desktop and jumbo. Leaving here as a reminder that it
+    // is an option.
+    // desktop: '(min-width: 992px) and (max-width: 1200px)',
+    // jumbo: '(min-width: 1201px)',
+};