+++ mirage/factories/root.ts
@@ -27,6 +27,7 @@ export default Factory.extend<Root>({
         ...Object.values(navigation),
         storageI18n,
         verifyEmailModals,
+        'ember_registries_new_style',
     ],
     message: 'Welcome to the OSF API.',
     version: '2.8',