+++ mirage/scenarios/default.ts
@@ -46,4 +46,7 @@ export default function(server: Server) {
         currentUserPermissions: Object.values(Permission),
     }, 'withRegisteredMeta');
     draftRegisterNodeMultiple(server, nodes[0], 12, {}, 'withRegistrationMetadata');
+    server.create('registration', {
+        id: 'decaf',
+    });
 }